adjusting for existing tables

+++ db/migrate/20100817192647_add_account_id_to_events.rb
@@ -1,13 +1,17 @@
 class AddAccountIdToEvents < ActiveRecord::Migration
   def self.up
     for klass in [:events, :people, :person_groups, :users, :pages, :menus, :articles, :article_categories, :features, :featurable_sections, :testimonials, :galleries, :images]
-      add_column klass, :account_id, :integer, :default => 1
+      if ActiveRecord.connection.tables.include?(klass.to_s.downcase.pluralize)
+        add_column klass, :account_id, :integer, :default => 1
+      end
     end
   end
 
   def self.down
     for klass in [:events, :people, :person_groups, :users, :pages, :menus, :articles, :article_categories, :features, :featurable_sections, :testimonials, :galleries, :images]
-      remove_column klass, :account_id
+      if ActiveRecord.connection.tables.include?(klass.to_s.downcase.pluralize)
+        remove_column klass, :account_id
+      end
     end
   end
 end
